perf(ui): lazy-load the upgrade modal body off the initial bundle
Split UpgradeModal.tsx into a tiny always-mounted gate (GlobalUpgradeModal)
and a lazily-imported content module (UpgradeModalContent). The modal's render
code is now fetched only when a plan-limit / seats / lite-member dialog opens,
instead of during initial page load.

Behavior is preserved: the gate returns null while closed and lazy-mounts the
same dialog on open (Suspense fallback={null}). Both production consumers
(DashboardLayout, OptimizationStudio) keep the same import path, so the
"mount on every full-screen surface" fix is intact.

Measured impact (rolldown prod build, before/after graph diff): defers the
modal's own ~6 kB raw / 2.4 kB gzip out of the dashboard eager closure. The
130 kB "UpgradeModal-*.js" chunk is a shared UI-commons chunk (60 importers),
NOT modal code, so it stays eager — the real bytes are tracked in Tier 2.

Tests: 25 pass (21 unit + 4 integration, incl. a new GlobalUpgradeModal gate
test covering closed=renders-nothing and open=lazy-loads-and-renders).
