+++ langwatch/src/stores/__tests__/upgradeModalStore.unit.test.ts
@@ -190,7 +190,9 @@ describe("upgradeModalStore", () => {
 
   describe("MODAL_CONTENT map", () => {
     it("has an entry for every variant mode", async () => {
-      const { MODAL_CONTENT } = await import("../../components/UpgradeModal");
+      const { MODAL_CONTENT } = await import(
+        "../../components/UpgradeModalContent"
+      );
       const expectedModes: Array<UpgradeModalVariant["mode"]> = [
         "limit",
         "seats",
